SISRP-46907 - Move HubEdos::V1 and HubEdos::V2 under HubEdos::StudentApi module
