+++ app/controllers/advising_student_controller.rb
@@ -54,7 +54,7 @@ class AdvisingStudentController < ApplicationController
     render json: {
       academicRoles: MyAcademics::MyAcademicRoles.new(student_uid).get_feed,
       attributes: User::AggregatedAttributes.new(student_uid).get_feed,
-      contacts: HubEdos::V1::Contacts.new(user_id: student_uid, include_fields: %w(names addresses phones emails)).get,
+      contacts: HubEdos::StudentApi::V1::Contacts.new(user_id: student_uid, include_fields: %w(names addresses phones emails)).get,
       residency: MyAcademics::Residency.new(student_uid).get_feed
     }
   end
