+++ app/models/campus_solutions/user_api_expiry.rb
@@ -1,10 +1,10 @@
 module CampusSolutions
   module UserApiExpiry
     def self.expire(uid=nil)
-      HubEdos::V1::MyStudent.expire uid
-      HubEdos::V1::Affiliations.expire uid
-      HubEdos::V1::Contacts.expire uid
-      HubEdos::V1::Demographics.expire uid
+      HubEdos::StudentApi::V1::MyStudent.expire uid
+      HubEdos::StudentApi::V1::Affiliations.expire uid
+      HubEdos::StudentApi::V1::Contacts.expire uid
+      HubEdos::StudentApi::V1::Demographics.expire uid
       User::Api.expire uid
     end
   end
