+++ app/models/campus_solutions/work_experience_updating_model.rb
@@ -3,7 +3,7 @@ module CampusSolutions
     def passthrough(model_name, params)
       proxy = model_name.new({user_id: @uid, params: params})
       result = proxy.get
-      HubEdos::V1::WorkExperience.expire @uid
+      HubEdos::StudentApi::V1::WorkExperience.expire @uid
       result
     end
   end
