+++ app/models/hub_edos/student_api/v1/gender.rb
@@ -1,0 +1,22 @@
+module HubEdos
+  module StudentApi
+    module V1
+      class Gender < Student
+        include HubEdos::CachedProxy
+        include Cache::UserCacheExpiry
+
+        def url
+          "#{@settings.base_url}/v1/students/#{@campus_solutions_id}/gender"
+        end
+
+        def json_filename
+          'hub_gender.json'
+        end
+
+        def whitelist_fields
+          %w(gender)
+        end
+      end
+    end
+  end
+end
