+++ spec/models/hub_edos/student_api/v1/demographics_spec.rb
@@ -1,8 +1,8 @@
-describe HubEdos::V1::Demographics do
+describe HubEdos::StudentApi::V1::Demographics do
 
   context 'mock proxy' do
     let(:include_fields) { nil }
-    let(:proxy) { HubEdos::V1::Demographics.new(fake: true, user_id: '61889', include_fields: include_fields) }
+    let(:proxy) { HubEdos::StudentApi::V1::Demographics.new(fake: true, user_id: '61889', include_fields: include_fields) }
     subject { proxy.get }
 
     it_behaves_like 'a proxy that properly observes the profile feature flag'
