+++ spec/models/hub_edos/student_api/v1/my_student_spec.rb
@@ -1,6 +1,6 @@
-describe HubEdos::V1::MyStudent do
+describe HubEdos::StudentApi::V1::MyStudent do
   subject {
-    proxy = HubEdos::V1::MyStudent.new(random_id, options)
+    proxy = HubEdos::StudentApi::V1::MyStudent.new(random_id, options)
     proxy.get_feed_internal
   }
   before do
