+++ spec/models/hub_edos/student_api/v1/registrations_spec.rb
@@ -1,7 +1,7 @@
-describe HubEdos::V1::Registrations do
+describe HubEdos::StudentApi::V1::Registrations do
 
   context 'mock proxy' do
-    let(:proxy) { HubEdos::V1::Registrations.new(fake: true, user_id: '61889') }
+    let(:proxy) { HubEdos::StudentApi::V1::Registrations.new(fake: true, user_id: '61889') }
     subject { proxy.get }
 
     it_should_behave_like 'a simple proxy that returns errors'
