+++ spec/models/hub_edos/student_api/v1/student_spec.rb
@@ -1,7 +1,7 @@
-describe HubEdos::V1::Student do
+describe HubEdos::StudentApi::V1::Student do
 
   context 'mock proxy' do
-    let(:proxy) { HubEdos::V1::Student.new(fake: true, user_id: '61889') }
+    let(:proxy) { HubEdos::StudentApi::V1::Student.new(fake: true, user_id: '61889') }
     subject { proxy.get }
 
     it_behaves_like 'a proxy that properly observes the profile feature flag'
