+++ spec/models/hub_edos/student_api/v1/work_experience_spec.rb
@@ -1,7 +1,7 @@
-describe HubEdos::V1::WorkExperience do
+describe HubEdos::StudentApi::V1::WorkExperience do
 
   context 'mock proxy' do
-    let(:proxy) { HubEdos::V1::WorkExperience.new(fake: true, user_id: '61889') }
+    let(:proxy) { HubEdos::StudentApi::V1::WorkExperience.new(fake: true, user_id: '61889') }
     subject { proxy.get }
 
     it_behaves_like 'a proxy that properly observes the profile feature flag'
