+++ spec/models/my_academics/residency_spec.rb
@@ -6,12 +6,12 @@ describe MyAcademics::Residency do
   let(:uid) { '61889' }
   # As deciphered by ResidencyMessageCode
   let(:message_nbr) { '2005' }
-  let(:fake_demographics_proxy) { HubEdos::V1::Demographics.new(fake: true, user_id: uid) }
+  let(:fake_demographics_proxy) { HubEdos::StudentApi::V1::Demographics.new(fake: true, user_id: uid) }
   let(:fake_residency_message_proxy) { CampusSolutions::ResidencyMessage.new(fake: true, messageNbr: message_nbr) }
 
   context 'fake data' do
     before do
-      allow(HubEdos::V1::Demographics).to receive(:new).with(user_id: uid).and_return fake_demographics_proxy
+      allow(HubEdos::StudentApi::V1::Demographics).to receive(:new).with(user_id: uid).and_return fake_demographics_proxy
       allow(CampusSolutions::ResidencyMessage).to receive(:new).with({messageNbr: message_nbr}).and_return fake_residency_message_proxy
     end
     it 'contains the expected data' do
